daemon/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-multipart>=0.0.12
psutil>=6.0.0
libvirt-python>=10.0.0
cryptography>=43.0.0
aiofiles>=24.1.0
httpx>=0.27.0
